requirements.txt
black==22.6.0
discord.py==2.0.0a4496+g584c22f6
pre-commit==2.20.0
